wip: qchdman: prepared code for later job-mode

+++ tools/qchdman/projectwindow.cpp
@@ -10,22 +10,27 @@
 extern Settings *globalConfig;
 extern MainWindow *mainWindow;
 
-ProjectWindow::ProjectWindow(QString pn, QWidget *parent) :
+ProjectWindow::ProjectWindow(QString pn, int type, QWidget *parent) :
     QMdiSubWindow(parent)
 {
     closeOk = true;
+    subWindowType = type;
 
     if ( pn.isEmpty() )
         projectName = tr("Noname-%1").arg(mainWindow->nextProjectID++);
     else
         projectName = pn;
 
-    projectWidget = new ProjectWidget(this);
-    setWidget(projectWidget);
-    setWindowTitle(projectName);
-    mainWindow->enableActions();
+    if ( subWindowType == QCHDMAN_MDI_PROJECT ) {
+        projectWidget = new ProjectWidget(this);
+        setWidget(projectWidget);
+        setWindowTitle(projectName);
+    } else if ( subWindowType == QCHDMAN_MDI_JOB ) {
+        // FIXME
+    }
 
     connect(this, SIGNAL(windowStateChanged(Qt::WindowStates, Qt::WindowStates)), this, SLOT(myWindowStateChanged(Qt::WindowStates, Qt::WindowStates)));
+    mainWindow->enableActions();
 }
 
 ProjectWindow::~ProjectWindow()
@@ -39,50 +44,58 @@ ProjectWindow::~ProjectWindow()
 
 void ProjectWindow::myWindowStateChanged(Qt::WindowStates oldState, Qt::WindowStates newState)
 {
-    if ( newState == Qt::WindowNoState && oldState == Qt::WindowActive)
-        return;
+    if ( subWindowType == QCHDMAN_MDI_PROJECT ) {
+        if ( newState == Qt::WindowNoState && oldState == Qt::WindowActive)
+            return;
 
-    if ( newState == Qt::WindowActive && oldState == Qt::WindowNoState )
-        return;
+        if ( newState == Qt::WindowActive && oldState == Qt::WindowNoState )
+            return;
 
-    projectWidget->on_comboBoxProjectType_currentIndexChanged(-1);
+        projectWidget->on_comboBoxProjectType_currentIndexChanged(-1);
+    } else if ( subWindowType == QCHDMAN_MDI_JOB ) {
+        // FIXME
+    }
 }
 
 void ProjectWindow::closeEvent(QCloseEvent *e)
 {
     closeOk = true;
 
-    if ( mainWindow->forceQuit ) {
-        projectWidget->chdmanProc->kill();
-        projectWidget->chdmanProc->waitForFinished();
-        e->accept();
-        deleteLater();
-        return;
-    }
+    if ( subWindowType == QCHDMAN_MDI_PROJECT ) {
+        if ( mainWindow->forceQuit ) {
+            projectWidget->chdmanProc->kill();
+            projectWidget->chdmanProc->waitForFinished();
+            e->accept();
+            deleteLater();
+            return;
+        }
 
-    if ( !mainWindow->closeOk ) {
-        e->ignore();
-        return;
-    }
+        if ( !mainWindow->closeOk ) {
+            e->ignore();
+            return;
+        }
 
-    if ( projectWidget->chdmanProc ) {
-        if ( projectWidget->chdmanProc->state() == QProcess::Running ) {
-            switch ( QMessageBox::question(this, tr("Confirm"),
-                                           tr("Project '%1' is currently running.\n\nClosing its window will kill the external process!\n\nProceed?").arg(windowTitle()),
-                                           QMessageBox::Yes | QMessageBox::No, QMessageBox::No) ) {
-            case QMessageBox::Yes:
-                projectWidget->chdmanProc->kill();
-                projectWidget->chdmanProc->waitForFinished();
-                break;
-            case QMessageBox::No:
-            default:
-                closeOk = false;
-                mainWindow->closeOk = false;
-                QTimer::singleShot(100, mainWindow, SLOT(resetCloseFlag()));
-                break;
-            }
-        } else
-            closeOk = mainWindow->closeOk;
+        if ( projectWidget->chdmanProc ) {
+            if ( projectWidget->chdmanProc->state() == QProcess::Running ) {
+                switch ( QMessageBox::question(this, tr("Confirm"),
+                                               tr("Project '%1' is currently running.\n\nClosing its window will kill the external process!\n\nProceed?").arg(windowTitle()),
+                                               QMessageBox::Yes | QMessageBox::No, QMessageBox::No) ) {
+                case QMessageBox::Yes:
+                    projectWidget->chdmanProc->kill();
+                    projectWidget->chdmanProc->waitForFinished();
+                    break;
+                case QMessageBox::No:
+                default:
+                    closeOk = false;
+                    mainWindow->closeOk = false;
+                    QTimer::singleShot(100, mainWindow, SLOT(resetCloseFlag()));
+                    break;
+                }
+            } else
+                closeOk = mainWindow->closeOk;
+        }
+    } else if ( subWindowType == QCHDMAN_MDI_JOB ) {
+        // FIXME
     }
 
     if ( closeOk ) {
